+++ lib/models/imagery/vector.dart
@@ -4,6 +4,7 @@ import 'package:every_door/models/imagery/vector/style_reader.dart';
 import 'package:every_door/models/plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map_tile_caching/flutter_map_tile_caching.dart';
+import 'package:http/io_client.dart';
 import 'package:logging/logging.dart';
 import 'package:vector_map_tiles/vector_map_tiles.dart';
 
@@ -35,7 +36,10 @@ class VectorImagery extends Imagery {
     super.best = false,
     String cachingStore = kTileCacheImagery,
   }) : _tileProvider = FMTCTileProvider(
-            stores: {cachingStore: BrowseStoreStrategy.readUpdateCreate});
+          stores: {cachingStore: BrowseStoreStrategy.readUpdateCreate},
+          headers: headers,
+          httpClient: IOClient(),
+        );
 
   @override
   Future<void> initialize() async {