+++ lib/screens/editor/building.dart
@@ -12,7 +12,8 @@ import 'package:every_door/screens/editor.dart';
 import 'package:flutter/material.dart';
 import 'package:every_door/constants.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:every_door/generated/l10n/app_localizations.dart' show AppLocalizations;
+import 'package:every_door/generated/l10n/app_localizations.dart'
+    show AppLocalizations;
 import 'package:latlong2/latlong.dart' show LatLng;
 
 class BuildingEditorPane extends ConsumerStatefulWidget {
@@ -372,8 +373,10 @@ class _BuildingEditorPaneState extends ConsumerState<BuildingEditorPane> {
                         Navigator.pushReplacement(
                           context,
                           MaterialPageRoute(
-                            builder: (context) =>
-                                PoiEditorPage(amenity: building),
+                            builder: (context) => PoiEditorPage(
+                              amenity: building,
+                              isModified: building != widget.building,
+                            ),
                             fullscreenDialog: true,
                           ),
                         );